+++ lyr-core/src/main/java/com/lyr/rule/dynamodb/config/ScanDynamodbTableIdleRuleConfigCreator.java
@@ -23,15 +23,25 @@ public class ScanDynamodbTableIdleRuleConfigCreator extends RuleConfigCreator<Sc
             return Optional.empty();
         }
 
-        final var configMap = (Map<String, Object>) input;
+        try {
+            final var configMap = (Map<String, Object>) input;
 
-        final var ruleConfigBuilder = ScanDynamodbTableIdleRuleConfig.builder();
-        RuleConfig.setConfigIfAttributePresent(
-                MAX_IDLE_PERIOD_IN_DAYS_CONFIG_KEY, configMap, ruleConfigBuilder::maxIdlePeriodInDays);
-        RuleConfig.setConfigIfAttributePresent(
-                EXCLUDE_EMPTY_TABLES_CONFIG_KEY, configMap, ruleConfigBuilder::excludeEmptyTables);
+            final var ruleConfigBuilder = ScanDynamodbTableIdleRuleConfig.builder();
+            RuleConfig.setConfigIfAttributePresent(
+                    MAX_IDLE_PERIOD_IN_DAYS_CONFIG_KEY, configMap, ruleConfigBuilder::maxIdlePeriodInDays);
+            RuleConfig.setConfigIfAttributePresent(
+                    EXCLUDE_EMPTY_TABLES_CONFIG_KEY, configMap, ruleConfigBuilder::excludeEmptyTables);
 
-        return Optional.of(ruleConfigBuilder.build());
+            return Optional.of(ruleConfigBuilder.build());
+        } catch (final Exception exception) {
+            log.atWarn()
+                    .addArgument(SCAN_DYNAMODB_TABLE_IDLE::getRuleName)
+                    .addArgument(exception.getClass().getName())
+                    .addArgument(exception.getMessage())
+                    .log(
+                            "Provided config for {} rule can not be parsed due to an error. Alternating to default config for this rule. ErrorType: {}, Error: {}");
+            return Optional.empty();
+        }
     }
 
     @Override
@@ -42,12 +52,12 @@ public class ScanDynamodbTableIdleRuleConfigCreator extends RuleConfigCreator<Sc
     @Override
     protected void validate(final ScanDynamodbTableIdleRuleConfig ruleConfig) {
         if (ruleConfig == null) {
-            throw new BadRuleConfigException("Rule config must not be null");
+            throw BadRuleConfigException.forNull(SCAN_DYNAMODB_TABLE_IDLE.getRuleName());
         }
 
         if (ruleConfig.getMaxIdlePeriodInDays() < 1) {
-            throw new BadRuleConfigException("Max idle period attribute of " + SCAN_DYNAMODB_TABLE_IDLE.getRuleName()
-                    + " rule config, must be longer than a day");
+            throw BadRuleConfigException.forLessThanLimit(
+                    MAX_IDLE_PERIOD_IN_DAYS_CONFIG_KEY, SCAN_DYNAMODB_TABLE_IDLE.getRuleName(), 1);
         }
     }
 }