+++ lyr-core/src/main/java/com/lyr/rule/glue/config/ScanGlueSessionActiveWithLongIdleTimeoutRuleConfigCreator.java
@@ -23,13 +23,23 @@ public class ScanGlueSessionActiveWithLongIdleTimeoutRuleConfigCreator
             return Optional.empty();
         }
 
-        final var configMap = (Map<String, Integer>) input;
+        try {
+            final var configMap = (Map<String, Integer>) input;
 
-        final var ruleConfigBuilder = ScanGlueSessionActiveWithLongIdleTimeoutRuleConfig.builder();
-        RuleConfig.setConfigIfAttributePresent(
-                MAX_IDLE_TIMEOUT_IN_MINUTES_CONFIG_KEY, configMap, ruleConfigBuilder::maxIdleTimeoutInMinutes);
+            final var ruleConfigBuilder = ScanGlueSessionActiveWithLongIdleTimeoutRuleConfig.builder();
+            RuleConfig.setConfigIfAttributePresent(
+                    MAX_IDLE_TIMEOUT_IN_MINUTES_CONFIG_KEY, configMap, ruleConfigBuilder::maxIdleTimeoutInMinutes);
 
-        return Optional.of(ruleConfigBuilder.build());
+            return Optional.of(ruleConfigBuilder.build());
+        } catch (final Exception exception) {
+            log.atWarn()
+                    .addArgument(SCAN_GLUE_SESSION_ACTIVE_WITH_LONG_IDLE_TIMEOUT::getRuleName)
+                    .addArgument(exception.getClass().getName())
+                    .addArgument(exception.getMessage())
+                    .log(
+                            "Provided config for {} rule can not be parsed due to an error. Alternating to default config for this rule. ErrorType: {}, Error: {}");
+            return Optional.empty();
+        }
     }
 
     @Override
@@ -40,13 +50,14 @@ public class ScanGlueSessionActiveWithLongIdleTimeoutRuleConfigCreator
     @Override
     protected void validate(final ScanGlueSessionActiveWithLongIdleTimeoutRuleConfig ruleConfig) {
         if (ruleConfig == null) {
-            throw new BadRuleConfigException("Rule config must not be null");
+            throw BadRuleConfigException.forNull(SCAN_GLUE_SESSION_ACTIVE_WITH_LONG_IDLE_TIMEOUT.getRuleName());
         }
 
         if (ruleConfig.getMaxIdleTimeoutInMinutes() < 0) {
-            throw new BadRuleConfigException("Max idle timeout in minutes attribute of "
-                    + SCAN_GLUE_SESSION_ACTIVE_WITH_LONG_IDLE_TIMEOUT.getRuleName()
-                    + " rule config, must be equal or greater than 0");
+            throw BadRuleConfigException.forLessThanLimit(
+                    MAX_IDLE_TIMEOUT_IN_MINUTES_CONFIG_KEY,
+                    SCAN_GLUE_SESSION_ACTIVE_WITH_LONG_IDLE_TIMEOUT.getRuleName(),
+                    0);
         }
     }
 }