+++ lyr-core/src/main/java/com/lyr/rule/lambda/config/ScanLambdaFunctionWithUnboundedConcurrencyRuleConfigCreator.java
@@ -1,5 +1,7 @@
 package com.lyr.rule.lambda.config;
 
+import static com.lyr.util.RuleDefinition.SCAN_LAMBDA_FUNCTION_WITH_UNBOUNDED_CONCURRENCY;
+
 import com.lyr.exception.rule.config.BadRuleConfigException;
 import com.lyr.rule.RuleConfigCreator;
 import java.util.Optional;
@@ -20,7 +22,7 @@ public class ScanLambdaFunctionWithUnboundedConcurrencyRuleConfigCreator
     @Override
     protected void validate(final ScanLambdaFunctionWithUnboundedConcurrencyRuleConfig ruleConfig) {
         if (ruleConfig == null) {
-            throw new BadRuleConfigException("Rule config must not be null");
+            throw BadRuleConfigException.forNull(SCAN_LAMBDA_FUNCTION_WITH_UNBOUNDED_CONCURRENCY.getRuleName());
         }
     }
 }