+++ lyr-core/src/test/java/com/lyr/cli/runners/ScanAwsEnvironmentRunnerTest.java
@@ -45,7 +45,7 @@ class ScanAwsEnvironmentRunnerTest {
     void testThatGivenNoUserDefinedRuleSetDefaultRulesAreCreatedAndEvaluatedAndAReportIsCreated() {
         // Given
         Settings.setAppSettings(Settings.builder().build());
-        final var numberOfDefaultRules = 4;
+        final var numberOfDefaultRules = 5;
         // When
         mockedRuleFactory.when(() -> RuleFactory.createRule(any())).thenReturn(mockedRule);
         mockedReporterFactory.when(() -> ReporterFactory.createReporter(any())).thenReturn(mockedReporter);
@@ -60,6 +60,9 @@ class ScanAwsEnvironmentRunnerTest {
                 () -> RuleFactory.createRule(RuleDefinition.SCAN_GLUE_SESSION_ACTIVE_WITH_LONG_IDLE_TIMEOUT), times(1));
         mockedRuleFactory.verify(
                 () -> RuleFactory.createRule(RuleDefinition.SCAN_CLOUDWATCH_LOG_GROUP_WITHOUT_RETENTION_POLICY),
+                times(1));
+        mockedRuleFactory.verify(
+                () -> RuleFactory.createRule(RuleDefinition.SCAN_LAMBDA_FUNCTION_WITH_DISALLOWED_ARCHITECTURE),
                 times(1));
         verify(mockedRule, times(numberOfDefaultRules)).evaluate();
         verify(mockedReporter, times(1)).report(any());